categorias

+++ Backend/src/modules/expenditure/services/DeleteExpenditureCategory.service.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from 'tsyringe';
+import AppError from '../../../shared/errors/AppError';
 import IExpenditureCategoryRepository from '../interfaces/IExpenditureCategoryRepository';
 
 @injectable()
@@ -8,7 +9,14 @@ export class DeleteExpenditureCategoryService {
         private repository: IExpenditureCategoryRepository
     ) { }
 
-    public async execute(id: number): Promise<void> {
+    public async execute(id: number, idCompany: number): Promise<void> {
+
+        const category = await this.repository.findByID(idCompany, id);
+
+        if (!category)
+            throw new AppError('Category not found')
+
+
         await this.repository.delete(id);
     }
